added gconnect, stuck at facebook login

+++ project.py
@@ -5,6 +5,9 @@ from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Restaurant, MenuItem, User
 
+#names the session login_session, which you will see throughout the code
+#i think its the same as just calling it session, but it probably refers to
+#something specific, ie the actual user's login session
 from flask import session as login_session
 import random, string
 
@@ -18,7 +21,7 @@ import requests
 CLIENT_ID = json.loads(
         open('client_secrets.json', 'r').read())['web']['client_id']
 #Connect to Database and create database session
-engine = create_engine('sqlite:///restaurantmenu.db')
+engine = create_engine('sqlite:///restaurantmenuwithusers.db')
 Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
@@ -26,10 +29,93 @@ session = DBSession()
 
 @app.route('/login/')
 def showLogin():
+    #randomly creates a hash or string 32 characters long from either 0-9 
+    #digits or uppercase ascii characters
     state = ''.join(random.choice(string.ascii_uppercase + string.
                 digits) for x in xrange(32))
+    #assigns the above state value to current login_session, used for
+    #verification later
     login_session['state'] = state
     return render_template('login.html', STATE=state)
+
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+    print "access token received as %s" % access_token
+    
+    #Exchange client token for long-lived server-side token with 
+    #GET /oauth/access_token?grant_type=fb_exchange_token&client_id={app-id}
+    #&client_secret={app-secret}%fb_exchange_token={short-lived-token}
+    print 'test'
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())['web']['app_id']                     
+    app_secret = json.loads(open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = ('https://graph.facebook.com/v2.9/oauth/access_token?'
+           'grant_type=fb_exchange_token&client_id=%s&client_secret=%s'
+           '&fb_exchange_token=%s') % (app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    print 'hello'
+    
+    #Use token to get user info from API    #long-term tokens can last up to two months
+    userinfo_url = "https://graph.facebook.com/v2.2/me"
+    #strip expire tag from access token since it's not need to make API calls
+    data = json.loads(result)
+    token = 'access_token=' + data['access_token']
+    
+    #if the token works, should be able to make API calls with new token like below:
+    url = 'https://graph.facebook.com/v2.9/me?%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    #populate login_session with user data from fb
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data['name']
+    login_session['email'] = data['email']
+    login_session['facebook_id'] = data['id']
+    
+    #Get user picture
+    url = 'https://graph.facebook.com/v2.9/me/picture?%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+    login_session['picture'] = data["data"]["url"]
+    
+    #see if user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+    
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    url = "https://graph.facebook.com/%s/permissions" % facebook_id
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+
+    del login_session['username']
+    del login_session['email']
+    del login_session['picture']
+    del login_session['user_id']
+    del login_session['facebook_id']
+    return "You have successfully logged out of Facebook."
+    
+#######    
 
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
@@ -89,6 +175,7 @@ def gconnect():
         return response
 
     # Store the access token in the session for later use.
+    login_session['provider'] = 'google'
     login_session['access_token'] = credentials.access_token
     login_session['gplus_id'] = gplus_id
 
@@ -128,30 +215,54 @@ def gdisconnect():
     print 'User name is: ' 
     print login_session['username']
     if access_token is None:
- 	print 'Access Token is None'
-    	response = make_response(json.dumps('Current user not connected.'), 401)
-    	response.headers['Content-Type'] = 'application/json'
-    	return response
+        print 'Access Token is None'
+        response = make_response(json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    
     url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print 'result is '
     print result
-    if result['status'] == '200':
-	del login_session['access_token'] 
-    	del login_session['gplus_id']
-    	del login_session['username']
-    	del login_session['email']
-    	del login_session['picture']
-    	response = make_response(json.dumps('Successfully disconnected.'), 200)
-    	response.headers['Content-Type'] = 'application/json'
-    	return response
-    else:
-	
-    	response = make_response(json.dumps('Failed to revoke token for given user.', 400))
-    	response.headers['Content-Type'] = 'application/json'
-    	return response
+#    if result['status'] == '200':
+#	del login_session['access_token'] 
+#    	del login_session['gplus_id']
+#    	del login_session['username']
+#    	del login_session['email']
+#    	del login_session['picture']
+#    	response = make_response(json.dumps('Successfully disconnected.'), 200)
+#    	response.headers['Content-Type'] = 'application/json'
+#    	return response
+#    else:
+#	
+#    	response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+#    	response.headers['Content-Type'] = 'application/json'
+#    	return response
 
+@app.route('/disconnect')
+def disconnect():
+    if 'provider' in login_session:
+        if login_session['provider'] == 'google':
+            gdisconnect()
+            del login_session['access_token'] 
+            del login_session['gplus_id']
+        if login_session['provider'] == 'facebook':
+            fbdisconnect()
+            del login_session['facebook_id']
+            
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        flash("You have successfully logged out.")
+        return redirect(url_for('showRestaurants'))
+    else:
+        flash("You were not logged in.")
+        print login_session['username']
+        return redirect(url_for('showRestaurants'))
+            
 #JSON APIs to view Restaurant Information
 @app.route('/restaurant/<int:restaurant_id>/menu/JSON')
 def restaurantMenuJSON(restaurant_id):
@@ -176,7 +287,10 @@ def restaurantsJSON():
 @app.route('/restaurant/')
 def showRestaurants():
   restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
-  return render_template('restaurants.html', restaurants = restaurants)
+  if 'username' not in login_session:
+      return render_template('publicrestaurants.html', restaurants = restaurants)
+  else: 
+      return render_template('restaurants.html', restaurants = restaurants)
 
 #Create a new restaurant
 @app.route('/restaurant/new/', methods=['GET','POST'])
@@ -214,6 +328,9 @@ def deleteRestaurant(restaurant_id):
   if 'username' not in login_session:
       return redirect('/login')
   restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+  if restaurantToDelete.user_id != login_session['user_id']:
+      return "<script>function myFunction() {alert('You are not allowed to \
+      delete this.  Please login.');}</script><body onload=myFunction()''>"
   if request.method == 'POST':
     session.delete(restaurantToDelete)
     flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -227,10 +344,15 @@ def deleteRestaurant(restaurant_id):
 @app.route('/restaurant/<int:restaurant_id>/menu/')
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    creator = getUserInfo(restaurant.user_id)
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
-    return render_template('menu.html', items = items, restaurant = restaurant)
+    if 'username' not in login_session or creator.id != login_session['user_id']:
+        return render_template('publicmenu.html', items = items, restaurant = 
+                               restaurant, creator = creator)
+    else:
+        return render_template('menu.html', items = items, restaurant = 
+                               restaurant, creator = creator)
      
-
 
 #Create a new menu item
 @app.route('/restaurant/<int:restaurant_id>/menu/new/',methods=['GET','POST'])
